add `setSwapBurner` calls + mulitple chains
